requirements.txt
Flask
reedsolo
io
pillow